+++ R/coercion.r
@@ -594,8 +594,8 @@ setMethod("as.character", signature(x = "Interval"), function(x, ...) {
 #'
 #' @section Compare to base R:
 #'
-#' These are drop in replacements for [as.Date()] and [as.POSIXct()],
-#' with a few tweaks to make them work more intuitively.
+#' These are drop in replacements for [as.Date()] and [as.POSIXct()], with a few
+#' tweaks to make them work more intuitively.
 #'
 #' \itemize{
 #'   \item `as_date()` ignores the timezone attribute, resulting in
@@ -608,10 +608,13 @@ setMethod("as.character", signature(x = "Interval"), function(x, ...) {
 #' @param x a vector of [POSIXt], numeric or character objects
 #' @param origin a Date object, or something which can be coerced by
 #'   `as.Date(origin, ...)` to such an object (default: the Unix epoch of
-#'   "1970-01-01"). Note that in this instance, `x` is assumed to reflect
-#'   the number of days since `origin` at `"UTC"`.
-#' @param tz a time zone name (default: time zone of the POSIXt object
-#'   `x`). See [OlsonNames()].
+#'   "1970-01-01"). Note that in this instance, `x` is assumed to reflect the
+#'   number of days since `origin` at `"UTC"`.
+#' @param tz a time zone name (default: time zone of the POSIXt object `x`). See
+#'   [OlsonNames()].
+#' @param format format argument for character methods. When supplied parsing is
+#'   performed by [strptime()]. For this reason consider using specialized
+#'   parsing functions in lubridate.
 #' @param ... further arguments to be passed to specific methods (see above).
 #' @return a vector of [Date] objects corresponding to `x`.
 #' @examples
@@ -644,8 +647,11 @@ setMethod(f = "as_date", signature = "numeric",
 #' @rdname as_date
 #' @export
 setMethod("as_date", "character",
-          function(x, tz = NULL) {
-            as_date(as_datetime(x, tz = "UTC"))
+          function(x, tz = NULL, format = NULL) {
+            if (is.null(format))
+              as_date(as_datetime(x, tz = "UTC"))
+            else
+              as_date(strptime(x, format, tz))
           })
 
 #' @rdname as_date
@@ -673,7 +679,12 @@ setMethod("as_datetime", "numeric",
 #' @rdname as_date
 #' @export
 setMethod("as_datetime", "character",
-          function(x, tz = "UTC") .parse_iso_dt(x, tz))
+          function(x, tz = "UTC", format = NULL) {
+            if (is.null(format))
+              .parse_iso_dt(x, tz)
+            else
+              strptime(x, format = format, tz = tz)
+          })
 
 #' @rdname as_date
 #' @export